Create useActivityLogLookupTables hook and move helpers out of ProjectActivityLog
